backend/db/db_client.py
import os
from supabase import create_client, Client

from dotenv import load_dotenv
load_dotenv()


SUPABASE_URL = os.environ.get("SUPABASE_URL")
SUPABASE_KEY = os.environ.get("SUPABASE_KEY")

if not SUPABASE_URL or not SUPABASE_KEY:
    raise RuntimeError("SUPABASE_URL and SUPABASE_KEY environment variables are required")

supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)


def log_conversation(record: dict):
    """Insert a conversation record into 'conversations' table."""
    supabase.table('conversations').insert(record).execute()