backend/requirements.txt
fastapi
uvicorn[standard]
langchain
google-cloud-aiplatform
openai
supabase
qdrant-client
sentence-transformers
python-dotenv
httpx
python-multipart